Add delete endpoints for author and magazine

+++ backend/src/main/java/org/mobelite/editormanager/repositories/MagazineRepository.java
@@ -5,12 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MagazineRepository extends JpaRepository<Magazine, Integer> {
+public interface MagazineRepository extends JpaRepository<Magazine, Long> {
 
     boolean existsMagazineByIssueNumber(int issueNumber);
-    
+
     List<Magazine> findMagazineByIssueNumber(int issueNumber);
 
     List<Magazine> findByTitleContainingIgnoreCase(String title);
-    
+
 }
